Splosions!
Added a pseudo-random explosion animation. There are still some kinks in it that can be found in the code for the 'explosions' function. This can be fixed by changing which way the leds are iterated over depending on if it is determined that the loop will overshoot the end of the array or the beginning of it.